+++ src/IpNetworkVector.h
@@ -69,6 +69,7 @@ public:
   LogicalVector isUnspecified() const;
   LogicalVector isLoopback() const;
   LogicalVector isLinkLocal() const;
+  LogicalVector isIPv4Mapped() const;
 };
 
 #endif